Fix all auth issues

+++ src/models/Categories/categories.actions.ts
@@ -1,14 +1,14 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import api from "../../services/endpoints/categories";
 import { toast } from "react-toastify";
-import { removeAuthToken } from "../../services/api";
 import { generateID } from "../../helpers/IdGenerator";
 import { logout } from "../Auth";
 import { fireCreateCategoryHook, fireDeleteCategoryHook } from "../Ui";
+import { ToastSuccess } from "src/coreUI/components/Typos/Alert";
 
 export const getAllCategories = createAsyncThunk<any, any, any>(
   "category/getAll",
-  async (args: any, { dispatch,rejectWithValue }) => {
+  async (args: any, { dispatch, rejectWithValue }) => {
     try {
       let result = await api.getCategories();
       if (result?.status === 401 || result?.status === 403) {
@@ -26,21 +26,16 @@ export const getAllCategories = createAsyncThunk<any, any, any>(
 
 export const createCategory = createAsyncThunk<any, any, any>(
   "category/createCategory",
-  async (args: any, { rejectWithValue }) => {
+  async (args: any, { dispatch, rejectWithValue }) => {
     try {
       let result = await api.createCategory(args.data);
+      if (result?.status === 401 || result?.status === 403) {
+        dispatch(logout(true));
+        return rejectWithValue("Un Authorized");
+      }
       if (result.data) {
-        args.dispatch(fireCreateCategoryHook(""))
-        toast.success("Category created successfully", {
-          position: "top-right",
-          autoClose: 1500,
-          hideProgressBar: false,
-          closeOnClick: true,
-          pauseOnHover: true,
-          draggable: true,
-          progress: undefined,
-          toastId:generateID(),
-        });
+        args.dispatch(fireCreateCategoryHook(""));
+        ToastSuccess("Category created successfully");
         return args;
       } else return [];
     } catch (error) {
@@ -51,20 +46,15 @@ export const createCategory = createAsyncThunk<any, any, any>(
 
 export const updateCategory = createAsyncThunk<any, any, any>(
   "category/updateCategory",
-  async (args: any, { rejectWithValue }) => {
+  async (args: any, { dispatch, rejectWithValue }) => {
     try {
       let result = await api.updateCategory(args);
+      if (result?.status === 401 || result?.status === 403) {
+        dispatch(logout(true));
+        return rejectWithValue("Un Authorized");
+      }
       if (result.data) {
-        toast.success("Category updated successfully", {
-          position: "top-right",
-          autoClose: 1500,
-          hideProgressBar: false,
-          closeOnClick: true,
-          pauseOnHover: true,
-          draggable: true,
-          progress: undefined,
-          toastId:generateID(),
-        });
+        ToastSuccess("Category updated successfully");
         return result.data;
       } else return [];
     } catch (error) {
@@ -75,21 +65,16 @@ export const updateCategory = createAsyncThunk<any, any, any>(
 
 export const deleteCategory = createAsyncThunk<any, any, any>(
   "category/deleteCategory",
-  async (args: any, { rejectWithValue }) => {
+  async (args: any, { dispatch, rejectWithValue }) => {
     try {
       let result = await api.updateCategory(args.data);
+      if (result?.status === 401 || result?.status === 403) {
+        rejectWithValue("Un Authorized");
+        return dispatch(logout(true));
+      }
       if (result.data) {
         args.dispatch(fireDeleteCategoryHook(""));
-        toast.success("Category deleted successfully", {
-          position: "top-right",
-          autoClose: 1500,
-          hideProgressBar: false,
-          closeOnClick: true,
-          pauseOnHover: true,
-          draggable: true,
-          progress: undefined,
-          toastId:generateID(),
-        });
+        ToastSuccess("Category deleted successfully");
         return result.data;
       } else return [];
     } catch (error) {
